+++ NEWS.md
@@ -1,3 +1,12 @@
+# versus 0.2.0
+
+* A new function `slice_diffs_both()` is added, used to get the differences from
+  both input tables in a single output with the rows interleaved
+
+* The error message for duplicated `by` values now shows the row which is duplicated
+
+* Comparisons are now faster
+
 # versus 0.1.0
 
 * Initial CRAN submission.